+++ src/main/lrphotos.lrdevplugin/Info.lua
@@ -4,7 +4,7 @@ return {
 
   LrPluginName = 'Photos',
   LrToolkitIdentifier = 'at.homebrew.lrphotos',
-  VERSION = { major = 2, minor = 0, revision = 0, build = 3, },
+  VERSION = { major = 2, minor = 1, revision = 0, build = 0, },
   LrInitPlugin = "InitPlugin.lua",
   -- Add the Metadata Definition File
   LrMetadataProvider = 'PhotosMetadataDefinition.lua',
@@ -26,7 +26,7 @@ return {
       file = "SearchExtraPhotosInPhotos.lua",
     },
     {
-      title = LOC "$$$/Photos/Menu/Library/MissingInLightroom=Search Photos in Lightroom with missing References",
+      title = LOC "$$$/Photos/Menu/Library/MissingInLightroom=Search missing Photos in Photos App",
       file = "SearchMissingPhotosInLightroom.lua",
     },
   },