+++ CHANGELOG.md
@@ -55,7 +55,7 @@ rendition is marked as skipped.
 
 ### Fixed
 
-* Fix reset issue for photosId tag.
+* Fix reset issue for photosId metadata field.
 
 ## [2.0.0.2] - 2024-02-01